+++ src/REST/Controller.php
@@ -9,8 +9,10 @@
 
 namespace FP\Privacy\REST;
 
+use FP\Privacy\Application\Consent\RevokeConsentHandler;
 use FP\Privacy\Consent\LogModel;
 use FP\Privacy\Frontend\ConsentState;
+use FP\Privacy\Presentation\REST\Controllers\ConsentController;
 use FP\Privacy\Utils\Options;
 use FP\Privacy\Admin\PolicyGenerator;
 
@@ -36,20 +38,19 @@ class Controller {
 	 */
 	public function __construct( ConsentState $state, Options $options, PolicyGenerator $generator, LogModel $log_model, $container = null ) {
 		$permission_checker = new RESTPermissionChecker();
-		// Use old handlers for now (backward compatibility).
-		// New controllers will be registered separately via RESTServiceProvider.
-		$summary_handler = new RESTSummaryHandler( $log_model, $options, $generator );
-		
-		// Try to get RevokeConsentHandler from container if available.
-		$revoke_handler = null;
-		if ( $container && method_exists( $container, 'has' ) && method_exists( $container, 'get' ) ) {
-			if ( $container->has( '\\FP\\Privacy\\Application\\Consent\\RevokeConsentHandler' ) ) {
-				$revoke_handler = $container->get( '\\FP\\Privacy\\Application\\Consent\\RevokeConsentHandler' );
+		$summary_handler    = new RESTSummaryHandler( $log_model, $options, $generator );
+
+		if ( $container && method_exists( $container, 'has' ) && $container->has( ConsentController::class ) ) {
+			$consent_handler = $container->get( ConsentController::class );
+		} else {
+			$revoke_handler = null;
+			if ( $container && method_exists( $container, 'has' ) && $container->has( RevokeConsentHandler::class ) ) {
+				$revoke_handler = $container->get( RevokeConsentHandler::class );
 			}
+			$consent_handler = new RESTConsentHandler( $state, $revoke_handler );
 		}
-		
-		$consent_handler    = new RESTConsentHandler( $state, $revoke_handler );
-		$revision_handler   = new RESTRevisionHandler( $options );
+
+		$revision_handler = new RESTRevisionHandler( $options );
 		$this->route_registrar = new RESTRouteRegistrar( $summary_handler, $consent_handler, $revision_handler, $permission_checker );
 	}